+++ documentation/changelog.md
@@ -1,3 +1,9 @@
 # Change Log
 <pre>
+v2.0	Exposed "trackLink" and "trackingServerSecure" [MOD-319]
+		BREAKING CHANGE: Android now returns a [Ti.Omniture.Session][] object from createSession, which has track and trackLink methods (for iOS parity)
+		Various Fixes for 1.8.0.1
+
 v1.0    Initial Release
+
+[Ti.Omniture.Session]: session.html